+++ src/main/java/com/example/task04app/external/openlibrary/config/OpenLibraryRequestInterceptor.java
@@ -29,15 +29,15 @@ public class OpenLibraryRequestInterceptor implements ClientHttpRequestIntercept
         logRequestDetails(request);
 
         ClientHttpResponse response = execution.execute(request, body);
-        log.debug("Response status open library: {}", response.getStatusCode().value());
+        log.trace("Response status open library: {}", response.getStatusCode().value());
 
         return response;
     }
 
     private void logRequestDetails(HttpRequest request) {
 
-        log.debug("Headers: {}", request.getHeaders());
-        log.debug("Request Method: {}", request.getMethod());
-        log.debug("Request URI: {}", request.getURI());
+        log.trace("Headers: {}", request.getHeaders());
+        log.trace("Request Method: {}", request.getMethod());
+        log.trace("Request URI: {}", request.getURI());
     }
 }
