+++ src/main/java/com/example/task04app/service/impl/BookServiceImpl.java
@@ -12,8 +12,8 @@ import com.example.task04app.exception.NotFoundException;
 import com.example.task04app.external.alfabank.dto.BookPriceInCurrenciesDto;
 import com.example.task04app.external.alfabank.pojo.Rate;
 import com.example.task04app.mapper.BookMapper;
-import com.example.task04app.entity.Book;
-import com.example.task04app.persistance.BookDao;
+import com.example.task04app.persistance.entity.Book;
+import com.example.task04app.service.persistance.BookDao;
 import com.example.task04app.service.BookService;
 import com.example.task04app.service.external.AlfaBankExchangeClient;
 import lombok.RequiredArgsConstructor;
